refactor(swiss-ai-hub): Promote ImapAgent out of playground into agents/
Addresses review feedback on #1622: the IMAP blueprint was the only agent
imported from playground/, forcing a special-case Dockerfile. Move it to the
standard location so it matches every other agent.

- Move ImapAgent, its config, start events, and BDD tests from
  playground/minimal_workflow/imap_workflow/ to
  swiss_ai_hub/agent/agents/imap_agent/ (configs/, events/, tests/)
- Move the manual trigger harness to app/imap_agent/trigger.py
- Register ImapAgent in agents/__init__.py (lazy export)
- Simplify the Dockerfile to the standard agent build (drop the three
  playground COPY lines)
- Fix alphabetical ordering of imap_agent in licenses.config.json and
  generate_compose.py (review nit)
- Update packages/agent/CLAUDE.md (Pre-Built Agents table, agents/ tree,
  remove imap_workflow from the playground list)
- Regenerate env-var reference doc (imap_agent's vars now documented)
