+++ infra/deployment/generate_compose.py
@@ -151,6 +151,7 @@ OWN_IMAGE_LICENSES = {
     "backup": "AGPL-3.0-or-later",
     "sysadmin-api": "AGPL-3.0-or-later",
     "sysadmin-web": "AGPL-3.0-or-later",
+    "imap_agent": "Apache-2.0",
     "llm_wrapping_agent": "Apache-2.0",
     "few_shot_agent": "Apache-2.0",
     "rag_agent": "Apache-2.0",
@@ -158,7 +159,6 @@ OWN_IMAGE_LICENSES = {
     "expert_asking_agent": "Apache-2.0",
     "namespace_selection_agent": "Apache-2.0",
     "memory_writer_agent": "Apache-2.0",
-    "imap_agent": "Apache-2.0",
     "retrieval_agent": "Apache-2.0",
     "default_rag_pipeline": "Apache-2.0",
     "shared_rag_pipeline": "Apache-2.0",
