+++ packages/agent/CLAUDE.md
@@ -16,6 +16,7 @@ packages/agent/                       # SDK framework
 │   │   ├── expert_asking_agent/       # Human expert escalation via Teams/Slack
 │   │   ├── expert_rag_agent/          # RAG with expert fallback
 │   │   ├── few_shot_agent/            # Pattern-matching with examples
+│   │   ├── imap_agent/                # Non-conversational IMAP mail read/move + draft-reply
 │   │   ├── namespace_selection_agent/ # LLM-driven knowledge routing
 │   │   └── retrieval_agent/           # Pure document retrieval (no LLM)
 │   ├── context/
@@ -348,6 +349,7 @@ the relevant steps.
 | **NamespaceSelectionAgent** | LLM-driven knowledge routing                              | HITL namespace approval + ThreadContext + RAG delegate                                                                                                                                                    |
 | **RetrievalAgent**          | Pure document retrieval (no LLM)                          | Retrieval-only, returns structured context                                                                                                                                                                |
 | **MemoryWriterAgent**       | System agent: async user-memory persistence (issue #1179) | Non-discoverable; triggered by `MemoryStorageRequestedEvent` from RAG/ExpertRAG when `enable_async_memory_storage` is on; rebuilds the origin agent's `AgentMemory` and writes off the chat critical path |
+| **ImapAgent**               | Non-conversational IMAP mail agent (read/move + draft-reply) | Two independent, programmatically-started chains: **read/move** (`ReadMailStartEvent` → list unread → fetch one message with S3-stored attachments → optionally move to the processed folder) and **draft** (`DraftMailStartEvent` → draft LLM replies for a batch of not-yet-drafted messages, `APPEND` to drafts, flag the source). Read-only IMAP (`BODY.PEEK`, source stays unread), at-least-once (append before flag), never sends. Grouped `ImapClientConfig`/`DraftEmailSettings` config in `packages/core/imap/` |
 
 Each agent has: `agents/{snake_name}/` (implementation), `app/{snake_name}/main.py` (entry point),
 `agents/{snake_name}/tests/` (BDD tests).
@@ -367,15 +369,7 @@ config seeder needed.
   `precondition_workflow`, `bounded_loop`, `context_workflow`, `configured_workflow`, `custom_start_stop_events`,
   `discoverable_workflow`, `displaying_workflow`, `multi_locale_workflow`, `optional_workflow`,
   `organization_memory_workflow`, `semantic_workflow`, `user_memory_workflow`, `multistep_human_in_the_loop_workflow`,
-  `long_running_agent`, `llama_index_workflow`, `mcp_react_workflow`, `imap_workflow` (non-conversational IMAP
-  demonstrator with two independent, separately-triggered chains — deployable via `app/imap_agent/main.py`):
-  **read/move** (`ReadMailStartEvent` → list unread → fetch one message with S3-referenced attachments → optionally move
-  it to the processed folder when `enable_move` is on → finish); and **draft** (`DraftMailStartEvent` → list up to
-  `batch_size` not-yet-drafted messages in `source_folder` (searched by `UNKEYWORD $AiHubDrafted`, or `UNANSWERED`
-  fallback) → for each: draft an LLM reply, `APPEND` it to the drafts folder, then flag the source message drafted →
-  finish). Reads use `BODY.PEEK` and marking never sets `\Seen`, so source mail stays unread; ordering is at-least-once
-  (append before flag); it never sends. The grouped `DraftEmailSettings` config (`packages/core/imap/`) holds the draft
-  toggle, source folder, batch size, drafts folder, LLM model, and prompt)
+  `long_running_agent`, `llama_index_workflow`, `mcp_react_workflow`
 - `playground/performance/` — Load testing with PerformanceTestingAgent
 
 ## Testing
