+++ packages/agent/app/imap_agent/main.py
@@ -7,8 +7,7 @@ import asyncio
 
 from swiss_ai_hub.core.infrastructure import AIHubSettings, enable_logging
 
-from playground.minimal_workflow.imap_workflow.imap_agent import ImapAgent
-from playground.minimal_workflow.imap_workflow.imap_agent_config import ImapAgentConfig
+from swiss_ai_hub.agent.agents.imap_agent import ImapAgent, ImapAgentConfig
 from swiss_ai_hub.agent.runners import AgentRunner
 
 enable_logging()
