+++ packages/agent/swiss_ai_hub/agent/agents/imap_agent/__init__.py
@@ -1,0 +1,4 @@
+from swiss_ai_hub.agent.agents.imap_agent.configs.imap_agent_config import ImapAgentConfig
+from swiss_ai_hub.agent.agents.imap_agent.imap_agent import ImapAgent
+
+__all__ = ["ImapAgent", "ImapAgentConfig"]
