+++ packages/agent/swiss_ai_hub/agent/agents/imap_agent/events/draft_mail_start_event.py
