+++ packages/agent/swiss_ai_hub/agent/agents/imap_agent/events/read_mail_start_event.py
