+++ packages/agent/swiss_ai_hub/agent/agents/imap_agent/tests/test_imap_agent.py
@@ -17,10 +17,10 @@ from swiss_ai_hub.core.i18n import LocaleString
 from swiss_ai_hub.core.imap import DraftEmailSettings, ImapClientConfig
 from swiss_ai_hub.core.testing import async_test
 
-from playground.minimal_workflow.imap_workflow.events.draft_mail_start_event import DraftMailStartEvent
-from playground.minimal_workflow.imap_workflow.events.read_mail_start_event import ReadMailStartEvent
-from playground.minimal_workflow.imap_workflow.imap_agent import ImapAgent
-from playground.minimal_workflow.imap_workflow.imap_agent_config import ImapAgentConfig
+from swiss_ai_hub.agent.agents.imap_agent.configs.imap_agent_config import ImapAgentConfig
+from swiss_ai_hub.agent.agents.imap_agent.events.draft_mail_start_event import DraftMailStartEvent
+from swiss_ai_hub.agent.agents.imap_agent.events.read_mail_start_event import ReadMailStartEvent
+from swiss_ai_hub.agent.agents.imap_agent.imap_agent import ImapAgent
 from swiss_ai_hub.agent.imap.parsed_message import ParsedAttachment, ParsedMessage
 from swiss_ai_hub.agent.runners import AgentTestRunner
